backend/requirements.txt
fastapi
uvicorn
python-dotenv
motor
pydantic
email-validator
resend
